MR-95: added form and subscriptions fixes
