+++ src/Components/Layout/index.js
@@ -10,6 +10,7 @@ import Home from '../Home/index';
 import Header from '../Header/index';
 import Footer from '../Footer/index';
 import styles from './layout.module.css';
+import FormSubscriptions from '../Subscriptions/FormSubscriptions/index';
 
 function Layout() {
   return (
@@ -23,9 +24,11 @@ function Layout() {
             <Route path="/admins" component={Admins} />
             <Route path="/classes" component={Classes} />
             <Route path="/members" component={Members} />
-            <Route path="/subscriptions" component={Subscriptions} />
+            <Route path="/subscriptions" exact component={Subscriptions} />
             <Route path="/superAdmins" component={SuperAdmins} />
             <Route path="/trainers" component={Trainers} />
+            <Route path="/subscriptions/form" exact component={FormSubscriptions} />
+            <Route path="/subscriptions/form/:id" component={FormSubscriptions} />
           </Switch>
         </div>
         <Footer />
